Run 'remove unnecessary cast' on Workspaces layer.

+++ src/Compilers/Core/Portable/FileSystem/PathUtilities.cs
@@ -685,7 +685,7 @@ namespace Roslyn.Utilities
                 {
                     if (!IsDirectorySeparator(ch))
                     {
-                        hc = Hash.Combine((int)char.ToUpperInvariant(ch), hc);
+                        hc = Hash.Combine(char.ToUpperInvariant(ch), hc);
                     }
                 }
             }
